refactor: split web constants.ts into colocated cluster modules
