+++ web/src/__tests__/utils/constants.property.test.ts
@@ -1,5 +1,5 @@
 import fc from 'fast-check'
-import { TASK_STATUS_ORDER, VALID_TRANSITIONS, TERMINAL_STATUSES } from '@/utils/constants'
+import { TASK_STATUS_ORDER, VALID_TRANSITIONS, TERMINAL_STATUSES } from '@/utils/tasks'
 
 describe('constants property tests', () => {
   const allStatuses = TASK_STATUS_ORDER
