+++ web/src/pages/SettingsNamespacePage.tsx
@@ -9,7 +9,7 @@ import { ErrorBoundary } from '@/components/ui/error-boundary'
 import { useSettingsStore } from '@/stores/settings'
 import { useSettingsData } from '@/hooks/useSettingsData'
 import { useSettingsDirtyState } from '@/hooks/useSettingsDirtyState'
-import { NAMESPACE_DISPLAY_NAMES, NAMESPACE_ORDER, SETTINGS_ADVANCED_KEY } from '@/utils/constants'
+import { NAMESPACE_DISPLAY_NAMES, NAMESPACE_ORDER, SETTINGS_ADVANCED_KEY } from '@/pages/settings/settings-constants'
 import { ROUTES } from '@/router/routes'
 import { FloatingSaveBar } from './settings/FloatingSaveBar'
 import { NamespaceSection } from './settings/NamespaceSection'
