+++ web/src/pages/SettingsPage.tsx
@@ -8,7 +8,7 @@ import { ErrorBanner } from '@/components/ui/error-banner'
 import { ErrorBoundary } from '@/components/ui/error-boundary'
 import { StaggerGroup, StaggerItem } from '@/components/ui/stagger-group'
 import { ToggleField } from '@/components/ui/toggle-field'
-import { NAMESPACE_DISPLAY_NAMES, NAMESPACE_ORDER } from '@/utils/constants'
+import { NAMESPACE_DISPLAY_NAMES, NAMESPACE_ORDER } from '@/pages/settings/settings-constants'
 import { AdvancedModeBanner } from './settings/AdvancedModeBanner'
 import { NotificationsSection } from './settings/NotificationsSection'
 import { SecurityConfigSection } from './settings/SecurityConfigSection'
