+++ web/src/pages/settings/SettingRow.tsx
@@ -2,7 +2,7 @@ import { useEffect, useId } from 'react'
 import { cn } from '@/lib/utils'
 import type { SettingEntry } from '@/api/types/settings'
 import { useFlash } from '@/hooks/useFlash'
-import { SECURITY_SENSITIVE_SETTINGS } from '@/utils/constants'
+import { SECURITY_SENSITIVE_SETTINGS } from '@/pages/settings/settings-constants'
 import { SourceBadge } from './SourceBadge'
 import { RestartBadge } from './RestartBadge'
 import { SettingField } from './SettingField'
