+++ web/src/pages/settings/ceremony-policy/CeremonyListPanel.tsx
@@ -3,7 +3,7 @@ import { ChevronDown, ChevronRight, List } from 'lucide-react'
 import type { CeremonyPolicyConfig, CeremonyStrategyType } from '@/api/types/ceremony-policy'
 import { InheritToggle } from '@/components/ui/inherit-toggle'
 import { SectionCard } from '@/components/ui/section-card'
-import { CEREMONY_STRATEGY_LABELS, STRATEGY_DEFAULT_VELOCITY_CALC } from '@/utils/constants'
+import { CEREMONY_STRATEGY_LABELS, STRATEGY_DEFAULT_VELOCITY_CALC } from '@/stores/ceremony-policy-constants'
 import { cn } from '@/lib/utils'
 import { StrategyPicker } from './StrategyPicker'
 import { StrategyConfigPanel } from './StrategyConfigPanel'
