+++ web/src/pages/settings/ceremony-policy/DepartmentOverridesPanel.tsx
@@ -5,7 +5,7 @@ import type { Department } from '@/api/types/org'
 import { InheritToggle } from '@/components/ui/inherit-toggle'
 import { SectionCard } from '@/components/ui/section-card'
 import { useCeremonyPolicyStore } from '@/stores/ceremony-policy'
-import { CEREMONY_STRATEGY_LABELS, STRATEGY_DEFAULT_VELOCITY_CALC } from '@/utils/constants'
+import { CEREMONY_STRATEGY_LABELS, STRATEGY_DEFAULT_VELOCITY_CALC } from '@/stores/ceremony-policy-constants'
 import { cn } from '@/lib/utils'
 import { StrategyPicker } from './StrategyPicker'
 import { StrategyConfigPanel } from './StrategyConfigPanel'
