+++ web/src/pages/workflows/WorkflowCreateDrawer.tsx
@@ -3,7 +3,7 @@ import { InputField } from '@/components/ui/input-field'
 import { SelectField } from '@/components/ui/select-field'
 import { SegmentedControl } from '@/components/ui/segmented-control'
 import { Button } from '@/components/ui/button'
-import { WORKFLOW_TYPES } from '@/utils/constants'
+import { WORKFLOW_TYPES } from '@/pages/workflows/workflow-constants'
 import { formatLabel } from '@/utils/format'
 import { BlueprintPicker } from './BlueprintPicker'
 import { useWorkflowCreateForm } from './useWorkflowCreateForm'
