+++ web/src/pages/workflows/workflow-constants.ts
@@ -1,0 +1,8 @@
+/** Workflow-creation constants. */
+
+export const WORKFLOW_TYPES = [
+  'sequential_pipeline',
+  'parallel_execution',
+  'kanban',
+  'agile_kanban',
+] as const
